refactor: snackbar error msg
