+++ src/hooks/useDeletePost.ts
@@ -3,6 +3,7 @@ import { usePathname, useRouter } from 'next/navigation.js';
 import { useAsyncFn } from 'react-use';
 
 import { type SocialSource } from '@/constants/enum.js';
+import { checkFarcasterInvalidSignerKey } from '@/helpers/checkers.js';
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
 import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.jsx';
 import { isRoutePathname } from '@/helpers/isRoutePathname.js';
@@ -29,6 +30,7 @@ export function useDeletePost(source: SocialSource) {
                 enqueueErrorMessage(getSnackbarMessageFromError(error, t`Failed to delete`), {
                     error,
                 });
+                checkFarcasterInvalidSignerKey(error);
                 throw error;
             }
         },
