+++ src/hooks/useReportPost.ts
@@ -2,6 +2,7 @@ import { t } from '@lingui/macro';
 import { useAsyncFn } from 'react-use';
 
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
+import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.jsx';
 import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
 
@@ -17,7 +18,9 @@ export function useReportPost() {
 
             return result;
         } catch (error) {
-            enqueueErrorMessage(t`Failed to submit report on ${post.source}`, { error });
+            enqueueErrorMessage(getSnackbarMessageFromError(error, t`Failed to submit report on ${post.source}`), {
+                error,
+            });
             throw error;
         }
     }, []);
