+++ src/hooks/useReportProfile.ts
@@ -2,6 +2,7 @@ import { t } from '@lingui/macro';
 import { useAsyncFn } from 'react-use';
 
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
+import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.jsx';
 import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
 import type { Profile } from '@/providers/types/SocialMedia.js';
 
@@ -15,10 +16,9 @@ export function useReportProfile() {
             const result = await provider.reportProfile(profile.profileId);
             if (result) enqueueSuccessMessage(t`Report submitted on ${profile.source}`);
             else enqueueErrorMessage(t`Failed to report @${profile.handle}`);
-
             return result;
         } catch (error) {
-            enqueueErrorMessage(t`Failed to report @${profile.handle}`, { error });
+            enqueueErrorMessage(getSnackbarMessageFromError(error, t`Failed to report @${profile.handle}`), { error });
             throw error;
         }
     }, []);
