+++ src/hooks/useToggleMutedChannel.ts
@@ -2,6 +2,7 @@ import { t } from '@lingui/macro';
 import { useAsyncFn } from 'react-use';
 
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
+import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.jsx';
 import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
 import type { Channel } from '@/providers/types/SocialMedia.js';
 
@@ -23,7 +24,10 @@ export function useToggleMutedChannel() {
             }
         } catch (error) {
             enqueueErrorMessage(
-                channel.blocked ? t`Failed to unmute /${channel.name}.` : t`Failed to mute /${channel.name}.`,
+                getSnackbarMessageFromError(
+                    error,
+                    channel.blocked ? t`Failed to unmute /${channel.name}.` : t`Failed to mute /${channel.name}.`,
+                ),
                 { error },
             );
             throw error;
