+++ src/hooks/useToggleMutedProfile.ts
@@ -2,6 +2,7 @@ import { t } from '@lingui/macro';
 import { useAsyncFn } from 'react-use';
 
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
+import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.jsx';
 import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
 import { resolveSourceName } from '@/helpers/resolveSourceName.js';
 import { useIsLogin } from '@/hooks/useIsLogin.js';
@@ -35,9 +36,12 @@ export function useToggleMutedProfile(operator: Profile | null) {
                 }
             } catch (error) {
                 enqueueErrorMessage(
-                    muted
-                        ? t`Failed to unmute @${profile.handle} on ${sourceName}`
-                        : t`Failed to mute @${profile.handle} on ${sourceName}`,
+                    getSnackbarMessageFromError(
+                        error,
+                        muted
+                            ? t`Failed to unmute @${profile.handle} on ${sourceName}`
+                            : t`Failed to mute @${profile.handle} on ${sourceName}`,
+                    ),
                     { error },
                 );
                 throw error;
